requirements.txt
fastapi
uvicorn
streamlit
requests
pandas
pydeck
groq
python-dotenv
pydantic